refactor(coverage-planning): consolidate algo section components
