+++ src/features/coverage-planning/components/AlgoSection/PathWidthInput/PathWidthInput.tsx
